[colors] move to colorsheet.py
