+++ visidata/colorsheet.py
